server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.109.2
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.19.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.5
langsmith==0.1.147
lazy_loader==0.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
openai==1.77.0
orjson==3.10.18
packaging==23.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.36.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.5.2
tokenizers==0.15.2
torch>=2.0.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.36.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.27.1
yarl==1.20.0
gTTS==2.5.1
diffusers==0.24.0
accelerate==0.25.0
Pillow>=9.0.0
